+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.2.9 (2025-12-25)
 
-- Add event markers for wait steps with SolversLib support. 
+- Add event markers for wait steps with SolversLib support.
 
 ## 1.2.8 (2025-12-23)